# Conflicts:
#	Modules/QtTools/QtTools/FileDialog/FileDialog.cpp
#	Modules/QtTools/QtTools/FileDialog/FileDialog.h
#	Modules/QtTools/QtTools/Utils/Utils.h
#	Projects/SceneViewer/Classes/FrameworkMain.cpp
#	Projects/TestBed/Classes/Tests/PackManagerTest.cpp
#	Sources/CMake/Modules/DavaTemplateModules.cmake
#	Sources/External/cef/CMakeLists.txt
#	Sources/Internal/Engine/Engine.h
#	Sources/Internal/Engine/Private/Engine.cpp
#	Sources/Internal/Engine/Private/EngineBackend.cpp
#	Sources/Internal/Engine/Private/Qt/RenderWidget.cpp
#	Sources/Internal/Engine/Qt/RenderWidget.h
#	Sources/Internal/UI/UILoadingScreen.h
#	Sources/Tools/QtTools/CMakeLists.txt
#	Sources/Tools/QtTools/FileDialog/FileDialog.cpp
#	Sources/Tools/QtTools/FileDialog/FileDialog.h
#	Sources/Tools/QtTools/FileDialogs/FileDialog.cpp
#	Sources/Tools/QtTools/FileDialogs/FileDialog.h
#	Tools/AssetCacheServer/Classes/UI/AssetCacheServerWindow.cpp
#	Tools/QuickEd/Classes/UI/FileSystemView/FileSystemDockWidget.cpp
#	Tools/ResourceEditor/Classes/Qt/Main/mainwindow.cpp
#	Tools/ResourceEditor/Classes/Qt/REApplication.cpp
#	Tools/ResourceEditor/Classes/Qt/ResourceEditorLauncher.h


Former-commit-id: e4cb3780d1976947e4ecdb6a209673d569c2cefe

+++ Sources/Internal/Render/2D/Sprite.cpp
@@ -816,8 +816,8 @@ void Sprite::DrawState::BuildStateFromParentAndLocal(const Sprite::DrawState& pa
         precomputedAngle = angle;
         if (precomputedAngle != parentState.angle)
         {
-            cosA = cosf(precomputedAngle);
-            sinA = sinf(precomputedAngle);
+            cosA = std::cos(precomputedAngle);
+            sinA = std::sin(precomputedAngle);
         }
         else
         {
